Moved the reducer into the store and created new store file
